+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### v0.3.11
 #### Added
 - Created a Troubleshooting page, comprising Diagnostics and Self-tests.
+#### Updated
+- Made the buttons, icons, and links on the catalog page visually consistent with the rest of the interface.
 
 ### v0.3.10
 #### Updated